changed delete behavior to check for active state and introduced new exception with new msg

+++ raumreservierung-backend/src/main/java/de/muenchen/raumreservierung/equipment/EquipmentService.java
@@ -1,15 +1,18 @@
 package de.muenchen.raumreservierung.equipment;
 
-import static de.muenchen.raumreservierung.common.ExceptionMessageConstants.MSG_NOT_FOUND;
-
+import de.muenchen.raumreservierung.common.ConflictException;
 import de.muenchen.raumreservierung.common.NotFoundException;
 import de.muenchen.raumreservierung.security.Authorities;
-import java.util.List;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.UUID;
+
+import static de.muenchen.raumreservierung.common.ExceptionMessageConstants.MSG_CANNOT_DELETE_ACTIVE;
+import static de.muenchen.raumreservierung.common.ExceptionMessageConstants.MSG_NOT_FOUND;
 
 @Service
 @Slf4j
@@ -32,8 +35,7 @@ public class EquipmentService {
 
     @PreAuthorize(Authorities.EQUIPMENT_MANAGE)
     public Equipment updateEquipment(final Equipment equipment, final UUID equipmentId) {
-        final Equipment foundEquipment = equipmentRepository.findById(equipmentId)
-                .orElseThrow(() -> new NotFoundException(String.format(MSG_NOT_FOUND, equipmentId)));
+        final Equipment foundEquipment = getEntityOrThrowException(equipmentId);
         foundEquipment.updateFrom(equipment);
         log.debug("Updating equipment to {}", foundEquipment);
         return equipmentRepository.save(foundEquipment);
@@ -41,8 +43,20 @@ public class EquipmentService {
 
     @PreAuthorize(Authorities.EQUIPMENT_MANAGE)
     public void deleteEquipment(final UUID equipmentId) {
+        final Equipment toDelete = getEntityOrThrowException(equipmentId);
+        
+        if (toDelete.isActive()) {
+            throw new ConflictException(String.format(MSG_CANNOT_DELETE_ACTIVE, equipmentId));
+        }
+
         log.debug("Deleting equipment {}", equipmentId);
         equipmentRepository.deleteById(equipmentId);
     }
 
+
+    private Equipment getEntityOrThrowException(final UUID equipmentId) {
+        return equipmentRepository
+                .findById(equipmentId)
+                .orElseThrow(() -> new NotFoundException(String.format(MSG_NOT_FOUND, equipmentId)));
+    }
 }
